refactor: 응집성을 고려하여 type 파일들의 폴더구조를 정리
